Cleanup
Adds some comments to remove warnings, and renames a few short variable
names in an attempt to make it slightly easier to read.

The loading of the CSV no longer parses the row before checking whether
the GlobalType means it should be saved, instead we just check the
specific index and bail early.